Fixed initialization of zero warnings counter.

+++ src/test/java/hudson/plugins/warnings/WarningsResultTest.java
@@ -24,7 +24,7 @@ public class WarningsResultTest extends AbstractAnnotationsBuildResultTest<Warni
 
     /** {@inheritDoc} */
     @Override
-    protected void verifyHighScoreMessage(final int expectedZeroWarningsBuildNumber, final boolean expectedIsNewHighScore, final long expectedHighScore, final int gap, final WarningsResult result) {
+    protected void verifyHighScoreMessage(final int expectedZeroWarningsBuildNumber, final boolean expectedIsNewHighScore, final long expectedHighScore, final long gap, final WarningsResult result) {
         if (result.hasNoAnnotations() && result.getDelta() == 0) {
             assertTrue(result.getDetails().contains(Messages.Warnings_ResultAction_NoWarningsSince(expectedZeroWarningsBuildNumber)));
             if (expectedIsNewHighScore) {
